+++ eslint/index.mjs
@@ -1,0 +1,1 @@
+export { default } from './defineEslintConfig.mjs';